DRY up records parsing in tools
