+++ shared/src/tools/generateCategories.js
@@ -8,6 +8,8 @@ const _ = require('lodash');
 const fs = require('fs');
 const parse = require('csv-parse');
 
+const { gatherRecords, getCsvOptions, sortableTitle } = require('./helpers');
+
 const USAGE = `
 Usage: node generateCategories.js [internal/external] spreadsheet.csv > output.json
 `;
@@ -75,24 +77,7 @@ if (type === 'internal') {
   ];
 }
 
-const csvOptions = {
-  columns: csvColumns,
-  delimiter: ',',
-  from_line: 2, // assumes first entry is header column containing labels
-  relax_column_count: true,
-  skip_empty_lines: true,
-  trim: true,
-};
-
-const whitespaceCleanup = str => {
-  str = str.replace(/[\r\n\t\s]+/g, ' ');
-  str = str.trim();
-  return str;
-};
-
-const sortableTitle = title => {
-  return whitespaceCleanup(title.toLowerCase());
-};
+const csvOptions = getCsvOptions(csvColumns);
 
 const documentTypeSort = (a, b) => {
   const [first, second] = [
@@ -152,17 +137,7 @@ const main = () => {
 
   const stream = parse(data, csvOptions);
 
-  const gatherRecords = function gatherRecords() {
-    let record;
-    while ((record = this.read())) {
-      record = _.pick(record, exportColumns);
-      Object.keys(record).forEach(key => {
-        record[key] = whitespaceCleanup(record[key]);
-      });
-      output.push(record);
-    }
-  };
-  stream.on('readable', gatherRecords);
+  stream.on('readable', gatherRecords(exportColumns, output));
   stream.on('end', () => {
     output.forEach(el => {
       if (el.category.length === 0) {
